Implement zooming in/out of the current Tab, by altering the visible points (which is already separated out from total sample history).
Also, needed to (and did, yay!) find a way to revert the revert which
made chart/logging intervals into integers instead of doubles as they
really should be.  Verified this works on MacOSX as well, which was
where the original problem was (basically, using static functions has
resolved this, instead of using static constants - still doen't make
sense that g++ on Mac got confused by this).

+++ src/chart/main.h
@@ -29,9 +29,9 @@
 
 typedef struct {
 	// Sampling
-	int chartDelta;
+	double chartDelta;
 	bool chartDeltaModified;
-	int loggerDelta;
+	double loggerDelta;
 	bool loggerDeltaModified;
 	int sampleHistory;
 	bool sampleHistoryModified;
